RDU2: Fix a bug in PIC probing code
Not all of the cdevs are guaranteed to have a non-NULL 'devname' fiels,
so check for that before trying to strcmp it.

Also there's no need to continue to loop over chardevs once a single PIC
deivce has been found. Add a break statememnt to terminate that loop as
soon as possible.