RDU2+: Add separate device tree for MX6+ variant
Add separate device tree for MX6+ variant so that it would be possible
to enable MX6+ specific functionality in PCIe driver.

For now this commit is less than ideal because it just make the PCIe
compatiblity change as a part of board's .dts file. The proper way to
fix this would be to include "imx6qp.dtsi", however this file is not
going to be avialible until upstream Barebox syncs up "dts" directory
with upstream Kernel.