commands/miitool: Allow specifying PHY by their full path
Make -s option to be a bit more consistent with how it is used in
other commands by allowing PHY device to be specified by either its
name of full path in /dev. This change also allows us to leverage
autocompletion since it will work as expected for paths in /dev.

Signed-off-by: Andrey Smirnov <andrew.smirnov@gmail.com>