net: phy: Add basic driver for MV88E6XXX switches from Marvell
Port a very abridged version of MV88E6XXX DSA driver from Linux
kernel. To simplify things, instead of bringing any of DSA
infrastructure from the kernel, we just create virtual PHY devices
capable of proper configuration for each of the ports.

This is useful for cases where some of the ports need to be configured
to provide Tx and/or Rx clock skews in RGMII.

Signed-off-by: Andrey Smirnov <andrew.smirnov@gmail.com>