memtest: report progress every 1% approx
The previous fixed 4K value would have been good for small memories,
but for larger ones it introduces a ton of overhead; especially if
we're notifying via RATP.

So dynamically compute how often we're going to report progress based
on the memory size tested. We set a lower limit of 4K, but then use
the first multiple of 2 below 1% of the total expected absolute
progress.