ARM: mmu: Share code for create_sections()
Code in mmu.c never creates anything but 1:1 mapping, and barring that
plus the call to __mmu_cache_flush(), the code in mmu-early.c's
version of the function is pretty much identical. To avoid code
duplication, move it to mmu.h and convert both mmu.c and mmu-early.c
to use it.

Signed-off-by: Andrey Smirnov <andrew.smirnov@gmail.com>