VFxxx: DCD: Remove read leveling and gate training delays
Read leveling delays are being specified as zero, so they are as good
as disabled and can be safely dropped. Gate training delay OTOH is
specified as 4/128 tCK for both data slices. Specifying it as
such for Data Byte 1, leads to that slice unusable* during POR startup
which is somehow is mitigated by double-reset hack in DCD.

Dropping gate training delays allows VF610 Tower board to sucessfully
PoR-boot without the need for double resetting of the DDRMC.

* The board fails to boot. When examined via JTAG in such a state
  only even bytes of DDR memory are functional.

Signed-off-by: Andrey Smirnov <andrew.smirnov@gmail.com>