usb: imx: implement support for limiting host to full speed
This is needed when host is known to not work properly in high speed
mode.

In linux, chipidea driver supports 'maximum-speed' device tree property.
When that is set to "full-speed", driver sets PFSC bit in PORTSC
register, which disallows use of high speed mode.

This patch implements same support for barebox.

Important technical detail is that PFSC bit is cleared by port reset,
this setting it has to be done in ehci->init() callback which is called
after ehci_reset().

Signed-off-by: Nikita Yushchenko <nikita.yoush@cogentembedded.com>