ARM: dts: rdu1: Add node corresponding to switch-connected EEPROM
Some versions of RDU1 come with a switch-attached EEPROM that is also
connected to RAVE SP via a I2C mux. On such devices is is possible to
access that EEPROM via a dedicated RAVE SP command. Add a Device Tree
node so a corresponding cdev would be created by the driver.

On RDU1's that do not expose switch-connected EEPROM that way this
change should be harmless and accesses to the device would result in
-EIO.

Signed-off-by: Andrey Smirnov <andrew.smirnov@gmail.com>