ARM: rdu: Introduce RDU part-number based fixup abstraction
Multiple generation of RDU boards from ZII come with slight hardware
variation that can be detected and fixed-up correctly based on part
number stored in EEPROMs that are accessible via RAVE SP.

In order to reduce common boilerplate between various RDU boards
introduce a bit of abstraction to allow declaring part-number specific
fixup in RDU board code.

TODO: If this is acceptable upstream this needs to be moved out of
RDU1 folder and shared with RDU2. RDU2's LCD type fixup needs to be
converted to using this instead of reading a LCD type NVMEM cell.

Signed-off-by: Andrey Smirnov <andrew.smirnov@gmail.com>