common: poller: Narrow down re-entrace protection
Narrow down re-entrace protection such that only poller function that
is currently running is guaranteed to not be called again. This should
be acceptable since majority of pollers are orthogonal to each other
and share no-resources.

This change is driven by desire to allow poller functions to call APIs
that themselves might be relying on other pollers to
function. Concrete example that inspired this change was RAVE SP LED
driver. That driver relies on SERDEV subsystem to provide low-level
transport. At the same time functions like led_blink() and led_flash()
rely on poller subsystem to function properly. All of this means we
need to be able to properly service SERDEV while being called from LED
subsystem poller function.

Signed-off-by: Andrey Smirnov <andrew.smirnov@gmail.com>