ARM: i.MX: Add max_load_size option to boards that will need it
The max_load_size option makes sure that only the portion of the image
is loaded to SRAM that fits into it.

Note this does not cover the whole available SRAM area for all boards,
for the bigger SRAM areas only a part is chosen that is enough to fit
the initial loader in.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>